BackEnd/requirements.txt
django==4.1
django-filter==22.1
djangorestframework==3.13.1
django-cors-headers==3.13.0
django-heroku==0.3.1
gunicorn==20.1.0
whitenoise==6.2.0
pymysql==1.0.2
Pillow==9.2.0
django-filter==22.1